Add basic CRUD handling for deployments
This is the big chunk of the application. It will be responsible for
managing the deployments in the database, validating the data posted by
the end user and serving responses to the end user of this API.

Adding the `uuid` package, for easier generation of deployment guids.

Adding validation functions to run through few simple checks of the
deployment and decide weather they are acceptable at that state.